🚚 move get and set state out of index
