+++ packages/atom.io/src/index.ts
@@ -3,6 +3,7 @@ import type { Store, Transceiver } from "atom.io/internal"
 import type { Json } from "atom.io/json"
 
 export * from "./atom"
+export * from "./get-set"
 export * from "./logger"
 export * from "./selector"
 export * from "./silo"
@@ -44,51 +45,6 @@ export type ReadonlySelectorToken<_> = {
 export type FamilyMetadata = {
 	key: string
 	subKey: string
-}
-
-export const capitalize = (str: string): string =>
-	str[0].toUpperCase() + str.slice(1)
-
-export const getState = <T>(
-	token: ReadonlySelectorToken<T> | StateToken<T>,
-	store: Store = IO.IMPLICIT.STORE,
-): T => {
-	const state =
-		IO.withdraw(token, store) ?? IO.withdrawNewFamilyMember(token, store)
-	if (state === null) {
-		throw new Error(
-			`${capitalize(token.type)} "${token.key}" not found in store "${
-				store.config.name
-			}".`,
-		)
-	}
-	return IO.getState__INTERNAL(state, store)
-}
-
-export const setState = <T, New extends T>(
-	token: StateToken<T>,
-	value: New | ((oldValue: T) => New),
-	store: Store = IO.IMPLICIT.STORE,
-): void => {
-	try {
-		IO.openOperation(token, store)
-	} catch (thrown) {
-		if (!(typeof thrown === `symbol`)) {
-			throw thrown
-		}
-		return
-	}
-	const state =
-		IO.withdraw(token, store) ?? IO.withdrawNewFamilyMember(token, store)
-	if (state === null) {
-		throw new Error(
-			`${capitalize(token.type)} "${token.key}" not found in store "${
-				store.config.name
-			}".`,
-		)
-	}
-	IO.setState__INTERNAL(state, value, store)
-	IO.closeOperation(store)
 }
 
 export const isDefault = (
